refactor: Member domain package 변경
