+++ src/main/java/com/hongikbros/jobmanager/member/domain/Member.java
@@ -1,4 +1,4 @@
-package com.hongikbros.jobmanager.member.domain.member;
+package com.hongikbros.jobmanager.member.domain;
 
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
