+++ src/test/java/com/hongikbros/jobmanager/common/fixture/member/MemberFixture.java
@@ -1,7 +1,7 @@
 package com.hongikbros.jobmanager.common.fixture.member;
 
 import com.hongikbros.jobmanager.common.utils.TestObjectUtils;
-import com.hongikbros.jobmanager.member.domain.member.Member;
+import com.hongikbros.jobmanager.member.domain.Member;
 
 public class MemberFixture {
     public static final Member MEMBER1 = TestObjectUtils.createMember(1L, 1L, "EunSeok",
